+++ src/extension/extension.ts
@@ -23,6 +23,7 @@ import { getRedHatService, TelemetryService } from "@redhat-developer/vscode-red
 import * as vscode from "vscode";
 import { KAOTO_FILE_PATH_GLOB } from "./helpers";
 import { VSCodeKaotoChannelApiProducer } from './../webview/VSCodeKaotoChannelApiProducer';
+import { HelpFeedbackProvider } from "../views/providers/HelpFeedbackProvider";
 
 let backendProxy: VsCodeBackendProxy;
 let telemetryService: TelemetryService;
@@ -52,7 +53,7 @@ export async function activate(context: vscode.ExtensionContext) {
     backendProxy: backendProxy,
   });
 
-  vscode.commands.registerCommand('kaoto.open.textualeditor', async() => {
+  vscode.commands.registerCommand('kaoto.open.textualeditor', async () => {
     if (kieEditorStore.activeEditor !== undefined) {
       const doc = await vscode.workspace.openTextDocument(kieEditorStore.activeEditor?.document.document.uri);
       await vscode.window.showTextDocument(doc, vscode.ViewColumn.Beside);
@@ -62,6 +63,11 @@ export async function activate(context: vscode.ExtensionContext) {
   vscode.commands.registerCommand('kaoto.open', (uri: vscode.Uri) => {
     vscode.commands.executeCommand('vscode.openWith', uri, 'webviewEditorsKaoto');
   });
+
+  /*
+   * register help & feedback view provider
+   */
+  context.subscriptions.push(vscode.window.registerTreeDataProvider('kaoto.help', new HelpFeedbackProvider()));
 
   const redhatService = await getRedHatService(context);
   telemetryService = await redhatService.getTelemetryService();